Update on the State-of-art
Update to the structure of the State-of-Art in order to make it a draft for the final report.
Also added some information regarding DNS which was missing.
To Do until delivery today:
Change the DOM analysis in the literature Review to only include articles and not Knowledge Base rules made by us.
Review 2.3 and 2.4